+++ fiftyone/core/odm/document.py
@@ -35,7 +35,11 @@ class SerializableDocument(object):
         return self.to_dict() == other.to_dict()
 
     def fancy_repr(
-        self, class_name=None, select_fields=None, exclude_fields=None
+        self,
+        class_name=None,
+        select_fields=None,
+        exclude_fields=None,
+        **kwargs
     ):
         """Generates a customizable string representation of the document.
 
@@ -57,13 +61,14 @@ class SerializableDocument(object):
                 )
             ):
                 continue
-
-            value = getattr(self, f)
-
-            if isinstance(value, ObjectId):
-                d[f] = str(value)
+            if f in kwargs:
+                d[f] = kwargs[f]
             else:
-                d[f] = value
+                value = getattr(self, f)
+                if isinstance(value, ObjectId):
+                    d[f] = str(value)
+                else:
+                    d[f] = value
 
         doc_name = class_name or self.__class__.__name__
         doc_str = fou.pformat(d)
@@ -136,6 +141,82 @@ class SerializableDocument(object):
         return cls.from_dict(d, extended=False)
 
 
+class SampleDocument(SerializableDocument):
+    """Interface for sample backing documents."""
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+    @property
+    def collection_name(self):
+        """The name of the MongoDB collection to which this sample belongs, or
+        ``None`` if it has not been added to a dataset.
+        """
+        return None
+
+    @property
+    def in_db(self):
+        """Whether the sample has been added to the database."""
+        return False
+
+    @property
+    def ingest_time(self):
+        """The time the sample was added to the database, or ``None`` if it
+        has not been added to the database.
+        """
+        return None
+
+    def has_field(self, field_name):
+        """Determines whether the sample has a field of the given name.
+
+        Args:
+            field_name: the field name
+
+        Returns:
+            True/False
+        """
+        raise NotImplementedError("Subclass must implement `has_field()`")
+
+    def get_field(self, field_name):
+        """Gets the field of the sample.
+
+        Args:
+            field_name: the field name
+
+        Returns:
+            the field value
+
+        Raises:
+            AttributeError: if the field does not exist
+        """
+        raise NotImplementedError("Subclass must implement `get_field()`")
+
+    def set_field(self, field_name, value, create=False):
+        """Sets the value of a field of the sample.
+
+        Args:
+            field_name: the field name
+            value: the field value
+            create (False): whether to create the field if it does not exist
+
+        Raises:
+            ValueError: if ``field_name`` is not an allowed field name or does
+                not exist and ``create == False``
+        """
+        raise NotImplementedError("Subclass must implement `set_field()`")
+
+    def clear_field(self, field_name):
+        """Clears the value of a field of the sample.
+
+        Args:
+            field_name: the field name
+
+        Raises:
+            ValueError: if the field does not exist
+        """
+        raise NotImplementedError("Subclass must implement `clear_field()`")
+
+
 class MongoEngineBaseDocument(SerializableDocument):
     """Mixin for all ``mongoengine.base.BaseDocument`` subclasses that
     implements the :class:`SerializableDocument` interface.
@@ -154,7 +235,9 @@ class MongoEngineBaseDocument(SerializableDocument):
         # pylint: disable=no-member
         return self._fields_ordered
 
-    def to_dict(self, extended=False):
+    def to_dict(
+        self, extended=False,
+    ):
         if extended:
             return json.loads(self._to_json())